Progress on Firebird implementation. Added test for null-value preservation.

+++ conserve/src/org/conserve/adapter/HsqldbAdapter.java
@@ -44,7 +44,7 @@ public class HsqldbAdapter extends AdapterBase
 	}
 
 	@Override
-	public String getVarCharKeyword(int length)
+	public String getVarCharKeyword(long length)
 	{
 		return getVarCharKeyword();
 	}
